perf: optimize PIN pad rendering
